Products with ID and common API Response

+++ ProductsInventory.Api/src/Data/Repositories/ProductsRepository.cs
@@ -1,3 +1,4 @@
+using Microsoft.EntityFrameworkCore;
 using ProductsInventory.Api.Data;
 using ProductsInventory.Api.Entities;
 
@@ -12,37 +13,81 @@ public class ProductsRepository : IProductsRepository
         _context = applicationDbContext;
     }
 
-    public Product Get(string id)
+    public async Task<Product> AddAsync(Product product)
     {
-        // var product = _context.products.Find(product => product.Id == id);
-        Product product = _context.products.Find(id);
+        await _context.products.AddAsync(product);
+        await _context.SaveChangesAsync();
         return product;
     }
 
-    public List<Product> GetAll()
+    public async Task<IEnumerable<Product>> GetProductsAsync()
     {
-        return _context.products.ToList();
+        return await _context.products.ToListAsync();
     }
 
-    public void Remove(string id)
+    public async Task DeleteAsync(Guid id)
     {
-        // var product = _context.products.Find(product => product.Id == id);
-        Product product = _context.products.Find(id);
-        // products.Remove(product);
+        var product = await _context.products.FindAsync(id);
+        if (product is null)
+        {
+            return;
+        }
+
         _context.products.Remove(product);
+        await _context.SaveChangesAsync();
+        return;
     }
 
-    public Product Save(Product product)
+
+
+    public async Task<Product> GetByIdAsync(Guid id)
     {
-        _context.products.Add(product);
-        _context.SaveChanges();
-        return product;
+        return await _context.products.FindAsync(id)!;
     }
 
-    public Product Update(Product product)
+    public async Task UpdateAsync(Product product)
     {
-        _context.products.Update(product);
-        _context.SaveChanges();
-        return product;
+        var existingProduct = await _context.products.FindAsync(product.Id);
+        if (existingProduct is null)
+        {
+            throw new KeyNotFoundException("Product not found");
+        }
+
+        _context.Entry(existingProduct).CurrentValues.SetValues(product);
+        await _context.SaveChangesAsync();
     }
+
+    // public Product Get(string id)
+    // {
+    //     // var product = _context.products.Find(product => product.Id == id);
+    //     Product product = _context.products.Find(id);
+    //     return product;
+    // }
+
+    // public List<Product> GetAll()
+    // {
+    //     return _context.products.ToList();
+    // }
+
+    // public void Remove(string id)
+    // {
+    //     // var product = _context.products.Find(product => product.Id == id);
+    //     Product product = _context.products.Find(id);
+    //     // products.Remove(product);
+    //     _context.products.Remove(product);
+    // }
+
+    // public Product Save(Product product)
+    // {
+    //     _context.products.Add(product);
+    //     _context.SaveChanges();
+    //     return product;
+    // }
+
+    // public Product Update(Product product)
+    // {
+    //     _context.products.Update(product);
+    //     _context.SaveChanges();
+    //     return product;
+    // }
 }
